requirements.txt
fastapi
uvicorn
python-dotenv
asyncpg
pydantic-settings
sqlalchemy
sqlmodel
alembic
psycopg2 
python-jose
pydantic[email]
python-multipart
redis
fastapi-mail 
itsdangerous
passlib[bcrypt]
celery
asgiref
eventlet
pytest-asyncio
httpx[http2]
aiosqlite
fakeredis[aioredis]